Update _config.yml
We've only begun to scratch the surface with Jekyll's blog aware functionality. This guide is only going to cover a couple more steps you might want to take for your blog.

You may have noticed that the URL of your blog post does not include the blog directory in it. In Jekyll we can control the structure of our permalinks by editing the _config.yml file we created in Step 8. So let's change our permalink structure to include the blog directory.

14 Edit the _config.yml file. Add the following line at the end of the file: